Factor out the declaration table section
