+++ frontend/src/components/NewBepiasViews/DeclarationsTableSection/utils.js
